refactor: payload moved to dto
