+++ src/main/java/com/softserveinc/dokazovi/dto/payload/SignUpRequest.java
@@ -1,4 +1,4 @@
-package com.softserveinc.dokazovi.entity.payload;
+package com.softserveinc.dokazovi.dto.payload;
 
 import lombok.Data;
 
@@ -7,7 +7,10 @@ import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.Size;
 
 @Data
-public class LoginRequest {
+public class SignUpRequest {
+
+	@NotBlank
+	private String name;
 
 	@NotBlank
 	@Email
@@ -16,5 +19,4 @@ public class LoginRequest {
 	@NotBlank
 	@Size(min = 3, max = 20, message = "Password length must be from 4 to 16 symbols")
 	private String password;
-
 }
