+++ src/main/java/com/softserveinc/dokazovi/service/UserService.java
@@ -3,7 +3,7 @@ package com.softserveinc.dokazovi.service;
 import com.softserveinc.dokazovi.dto.user.UserDTO;
 import com.softserveinc.dokazovi.entity.UserEntity;
 import com.softserveinc.dokazovi.entity.VerificationToken;
-import com.softserveinc.dokazovi.entity.payload.SignUpRequest;
+import com.softserveinc.dokazovi.dto.payload.SignUpRequest;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
