+++ src/test/java/com/softserveinc/dokazovi/util/StringToNameParserTest.java
@@ -1,7 +1,7 @@
 package com.softserveinc.dokazovi.util;
 
 import com.softserveinc.dokazovi.entity.UserEntity;
-import com.softserveinc.dokazovi.entity.payload.SignUpRequest;
+import com.softserveinc.dokazovi.dto.payload.SignUpRequest;
 import com.softserveinc.dokazovi.security.oauth2.user.OAuth2UserInfo;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
